+++ packages/appengine-contracts/src/app.ts
@@ -1,7 +1,7 @@
-import { AppDocument, AppDocumentSpec } from '@provisioner/contracts'
+import { KubeDocument, keyValue } from '@c6o/kubeclient-contracts'
+import { AppDocumentLabels, AppDocumentSpec } from '@provisioner/contracts'
 import { Options as generatorOptions } from 'generate-password'
 import { Flow } from './flow'
-import { keyValue } from './keyValue'
 
 export interface ServicePort {
     name?: string
@@ -68,7 +68,7 @@ export interface AppEngineAppSpec extends AppDocumentSpec {
     provisioner?: AppEngineAppSpecProvisioner
 }
 
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface AppEngineAppDocument extends KubeDocument<AppDocumentLabels, keyValue, AppEngineAppSpec> {
 
-export interface AppEngineAppDocument extends AppDocument {
-    spec: AppEngineAppSpec
 }