+++ packages/appengine-contracts/src/flow.ts
@@ -1,5 +1,5 @@
 import { Options as generatorOptions } from 'generate-password'
-import { keyValue } from './keyValue'
+import { keyValue } from '@c6o/kubeclient-contracts'
 export { Options as generatorOptions } from 'generate-password'
 
 export type functionExpression = string | Function // this is an inline function that's turned into an expression
@@ -30,7 +30,7 @@ export interface c6oExtensions {
     step?: number
     hasControls?: boolean
     errorMessage?: string
-    dataSource: string
+    dataSource?: string
 }
 
 export interface InquirePrompt {