+++ packages/appengine/src/templates/secrets.ts
@@ -1,14 +1,13 @@
-import { keyValue } from '@provisioner/appengine-contracts'
+import { keyValue } from '@c6o/kubeclient-contracts'
 
-export function getSecretTemplate(name: string, namespace: string, data: keyValue, labels?: keyValue ) {
+export function getSecretTemplate(name: string, namespace: string, data: keyValue) {
 
     return {
         apiVersion: 'v1',
         kind: 'Secret',
         metadata: {
             name: `${name}-secret`,
-            namespace: namespace,
-            labels
+            namespace: namespace
         },
         type: 'Opaque',
         data