+++ packages/appengine/src/templates/services.ts
@@ -1,14 +1,13 @@
 
-import { keyValue, ServicePort } from '@provisioner/appengine-contracts'
+import { ServicePort } from '@provisioner/appengine-contracts'
 
-export function getServiceTemplate(name: string, namespace: string, ports: ServicePort[], labels?: keyValue) {
+export function getServiceTemplate(name: string, namespace: string, ports: ServicePort[]) {
     return {
         apiVersion: 'v1',
         kind: 'Service',
         metadata: {
             name,
-            namespace: namespace,
-            labels
+            namespace: namespace
         },
         spec: {
             type: 'NodePort',