+++ packages/appengine/src/templates/volumes.ts
@@ -1,14 +1,14 @@
-import { Volume, keyValue } from '@provisioner/appengine-contracts'
+import { keyValue } from '@c6o/kubeclient-contracts'
+import { Volume } from '@provisioner/appengine-contracts'
 
-export function getPVCTemplate(volume: Volume, namespace: string, labels?: keyValue) {
+export function getPVCTemplate(volume: Volume, namespace: string) {
 
     return {
         kind: 'PersistentVolumeClaim',
         apiVersion: 'v1',
         metadata: {
             name: volume.name.toLowerCase(),
-            namespace: namespace,
-            labels
+            namespace: namespace
         },
         spec: {
             accessModes: [