+++ packages/c6o-system/src/mixins/api/host.ts
@@ -2,20 +2,19 @@ import { baseProvisionerType } from '../../'
 
 export const hostApiMixin = (base: baseProvisionerType) => class extends base {
 
-    async getApplicationFQDN(appName: string, namespace: string) {
-        // TODO: Make this not require a namespace
+    async getSystemFQDN() {
         const result = await this.manager.cluster.read(this.systemServerConfigMap('c6o-system'))
-        if (result.error) {
-            // TODO: log failure
-            return void 0
-        }
+        if (result.error)
+            throw result.error
 
         const host = result.object?.data?.HOST
+        if (!host)
+            throw new Error('HOST not found')
+        return host
+    }
 
-        if (!host) {
-            // TODO: log missing host
-            return void 0
-        }
+    async getApplicationFQDN(appName: string, namespace: string) {
+        const host = this.getSystemFQDN()
 
         const provisioner = await this.manager.getAppProvisioner('istio', 'istio-system')
         const prefix = await provisioner.getApplicationPrefix(appName, namespace)