+++ packages/c6o-system/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.27](https://github.com/c6o/provisioners/compare/v0.0.26...v0.0.27) (2021-02-03)
+
+
+### ✨ Features
+
+* **c6o-system:** Ensure default namespace is managed by c6o ([#151](https://github.com/c6o/provisioners/issues/151)) ([88fe2de](https://github.com/c6o/provisioners/commit/88fe2deb0be898aa073a83eaa1b6ba0a29dc215d))
+* **c6o-system:** Initial release of Harbour Master ([#154](https://github.com/c6o/provisioners/issues/154)) ([a500e7f](https://github.com/c6o/provisioners/commit/a500e7f5a0b9232ab7d6c58308b2280d7cdde1b2)), closes [#155](https://github.com/c6o/provisioners/issues/155) [#1239](https://github.com/c6o/provisioners/issues/1239)
+
+
+### 🐛 Bug Fixes
+
+* **istio:** Removed http routing from virtual service ([#184](https://github.com/c6o/provisioners/issues/184)) ([ab0c33e](https://github.com/c6o/provisioners/commit/ab0c33e74618d0c8a871fadec5ebc3829bb0495d))
+* Removed reference to incorrect AppObject/AppDocument ([f988ede](https://github.com/c6o/provisioners/commit/f988ede1d5f38c632900169c11ead51a1af1c500))
+
+
+### 📦 Code Refactoring
+
+* **appengine:** Using contracts and using inquire flow in cli and UI ([#156](https://github.com/c6o/provisioners/issues/156)) ([4d612b9](https://github.com/c6o/provisioners/commit/4d612b909ac4eaa0ecddf3355363e7429e517204))
+
+
+### ♻️ Chores
+
+* **c6o-system:** Cleaned up host API ([0e00b11](https://github.com/c6o/provisioners/commit/0e00b114a0018f40ca57a97ab3646d5444dc0fae))
+* Code formatting changes only! ([2df4e69](https://github.com/c6o/provisioners/commit/2df4e69ed0b6050e0f6f9ddbf8400273e7c8eff0))
+* update Vaadin to 14.4.5 ([fa3cdde](https://github.com/c6o/provisioners/commit/fa3cddead8be2233dcf6960bb0191b535510e35f))
+
+
+
+
+
 ## [0.0.26](https://github.com/c6o/provisioners/compare/v0.0.25...v0.0.26) (2021-01-01)